+++ changelog.md
@@ -1,8 +1,7 @@
-# 0.4.0.0
+# 0.3.4.0
 
 * Add `foldr` catamorphism and `fromInfinite` / `toInfinite` conversions.
 * Add `iterateWithIndex` and `iterateWithIndexM`.
-* Remove instances `Foldable` and `Traversable`, they are too dangerous to diverge.
 
 # 0.3.3.0